InfoBoxes/Content/Weather: Add decimal digit for rel. humidity
%d resolves less than 7 bits, while most sensors offer at least a
resolution of 10 bits.

+++ src/InfoBoxes/Content/Weather.cpp
@@ -47,7 +47,7 @@ UpdateInfoBoxHumidity(InfoBoxData &data)
   }
 
   // Set Value
-  data.UnsafeFormatValue( _T("%d"), (int)basic.humidity);
+  data.UnsafeFormatValue( _T("%2.1f"), (float)basic.humidity);
 }
 
 void
